bugfix: Http Header Range field parsing
Range field was not getting parsed right, since it wont have a 'size'
field like Content-Range. The code was parsing a Range field looking
like 'Range: bytes=from-to/size', but it actually looks like 'Range:
bytes=from-to'.
The 'to' field may be missing. If thats the case the generated Response
will specify a size of 0, which is a bug. Added to todo list.